Version 3.1.0
FossilOrigin-Name: f18e1d074892acae0f939ae5d830a29450c317ea58b02641ceba89b23e2cb719

+++ lib/field_types.cpp
@@ -35,10 +35,10 @@ namespace mysqlpp {
 
 void FieldTypes::init(const ResultBase* res)
 {
-	int num = res->num_fields();
+	size_t num = res->num_fields();
 	reserve(num);
-	for (int i = 0; i < num; i++) {
-		push_back(res->field(i).type());
+	for (size_t i = 0; i < num; i++) {
+		push_back(res->field(unsigned(i)).type());
 	}
 }
